refactor(routing): organize tests into unit/ subdirectory
Move existing test files into tests/unit/ to establish a clear test
hierarchy (unit, integration, container) for the routing package.
